server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
transformers==4.45.0
pillow
numpy
